Move packages to be product-centric not cloud-centric.
This change moves files to be product-centric rather than cloud-centric.
The issue with cloud-centric naming is that there can be multiple
products that do the same thing, e.g. AWS has both S3 and Glacier as
blob stores. The old naming scheme meant it was not possible to
differentiate between the two products. This change fixes that.

Change-Id: I8652a7dc2c650df8f04312a70f2202372c8d175b
Reviewed-on: https://team-review.git.corp.google.com/253315
Reviewed-by: Ross Light <light@google.com>
Reviewed-by: kokoro <noreply+kokoro@google.com>