Initial read of a paramstore parameter.
This change reads a manually inserted paramstore parameter and checks it
was successfully read.

The change is very preliminary and being sent out only to ensure it's
heading along the right track. In particular, the test should write it's
own parameters and the code itself doesn't conform to the runtimeconfig
interfaces.

Change-Id: I335437ded8edf652e4b53f3dcd8107855f057c81
Reviewed-on: https://team-review.git.corp.google.com/255111
Reviewed-by: Tuo Shan <shantuo@google.com>
Reviewed-by: Ross Light <light@google.com>