Add ability to watch for paramstore changes
This change adds a Watch function (implementing the Watcher
interface) that will block waiting for a particular parameter to change
in the paramstore.

Unfortunately AWS returns the desired data in two different API calls,
so I had to do that. It's unfortunate, but such is the life of API callers.

Change-Id: Iacc5b8022e0c52998d08afd417a76accd3f6abd3
Reviewed-on: https://team-review.git.corp.google.com/256516
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Tuo Shan <shantuo@google.com>
Reviewed-by: Ross Light <light@google.com>