Add a Kokoro check for the copyright notice
This change adds a quick grep to Kokoro to search for source files that
should have a copyright notice that don't. Cross would be better (as
described on the opensource.google.com docs), but it's not released so
would have to be run from google3, which we can't do automatically from
Kokoro.

This change also fixes those files that didn't have copyright notices by
accident (i.e. those I wrote :/).

Change-Id: Ib623fb98bc96012e8c18a75b71c28c1f44296363
Reviewed-on: https://team-review.git.corp.google.com/256971
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Ross Light <light@google.com>