paramstore: Modify Watch to tick immediately
Tickers don't actually tick immediately. The rsc@ advocated method is to
write a function that you call first: https://github.com/golang/go/issues/17601

This change modifies Watch so that it utilizes this strategy and will
return immediately. Checking canceled contexts is now a bit more fiddly
in the test, but not too bad.

Change-Id: Ia41e77400a1e47605459aaa56f9072b19aa643a6
Reviewed-on: https://team-review.git.corp.google.com/261051
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Tuo Shan <shantuo@google.com>