Stop touching replay files if not recording
Replay files are being touched for header fixing when no recording is
taking place. This means that runs are full of junk changes to the
replay files that don't mean anything. This change fixes that problem by
only fixing headers if the test is being run in a recording mode.

Fixes 38