wire: don't fail at first error
This changes both Generate and Load to try to advance as far as
possible while collecting errors.

To support this, I bundle errors together into a list that can be
returned into a single error, with the ability to unbundle them later.
The exported API surface only sees the unbundled errors so that each
can be formatted individually.

Fixes #5