wire: add error string assertions in tests
out.txts that start with "ERROR" can now include substrings to assert
exist in the error messages, one per line.  This gives a very
coarse-grained way of testing for error message quality. It is not a
substitute for manual verification, but gives a "good enough" sanity
check that relevant details are reported.

Updates #5