runtimevar/filevar: synchronize concurrent tests
These tests were not properly handling partial reads and were
using sleeps. This should make the tests faster and more robust.
Since the flake happens so infrequently, it's hard for me to claim
that I've definitively fixed it, but I'm unable to induce a
failure locally after this change.

Fixes #58