wire: use go/packages for analysis
Unfortunately, this does come with a ~4x slowdown to Wire, as it is
now pulling source for all transitively depended packages, but not
trimming comments or function bodies. This is due to limitations with
the ParseFile callback in go/packages.

This comes with a single semantic change: when performing analysis, Wire
will now evaluate everything with the wireinject build tag. I updated
the build tags tests accordingly.

I deleted the vendoring test, as vendoring becomes obsolete with
modules.

go/packages now parses comments by default, so now the generated code
includes comments for non-injector declarations.

Fixes #78