internal/proxy: add step to copy from GCS
Instead of starting from zero, instruct proxy updaters to start from the
current module cache by copying from the remote bucket. In common cases,
this should reduce the probability of accidental removal of dependencies
that other collaborators depend on in in-flight PRs.

There is still the possibility that someone else runs these steps
concurrently and clobbers work, but that seems much more remote, and the
resolution would be to go through the process again.

Fixes #661