wire: remove code generator tool
The marker function package remains where it is, but now aliases to the
new Wire package. I will update the import paths in this repository
in a separate PR so as to shake out any unforeseen issues that would be
encountered by end-users as well.

I removed the content of the README and left behind a pointer to the new
repository at https://github.com/google/wire. I fixed any dangling
documentation links I could find.

Updates #828