docs: initial page for our GitHub Pages site.
We plan to use the GitHub Pages site associated with this repo to
host redirection pages for our gocloud.dev vanity import path.

As a first step, we set up a trivial site to make sure the basic
Pages config is working.

Our site will live in the docs folder of the master branch.