internal/contributebot: obtain title patterns from new ".contributebot" file
Both the regular expression and the text of the response can be changed
using this new configuration file, which is plain JSON. I have added an
in-memory cache so as to avoid eating into the GitHub API quota for
every event.

I have tested the integration manually on my own test repository.

Updates google/wire#82