rabbitpubsub: add an in-memory fake
In the main rabbit code, replace the amqp client's types with interfaces
for AMQP's Connection and Channel resources.

Implement the interfaces by calling the actual amqp client. Use that
implementation for normal execution.

Write an in-memory fake that satisfies the same interfaces.

Use it for tests when a real RabbitMQ server is not available.

Fixes #837.