internal/admin: codify project configuration into Terraform (#999)
This allows us to version control and review changes to our repository
configuration just like other changes, as well as keep both Wire and Go
Cloud mostly in sync in terms of setup. Testing this already corrected
multiple such discrepancies between the projects, so this seems like a
good thing to have.