rabbitpubsub: collect all errors from SendBatch (#993)
Instead of just returning the first error we see when publishing, we collect
and return them all.

It doesn't make sense to do this for ReceiveBatch: we want to stop on the first error.

SendAcks is debatable. I don't think it's worth acking the remaining
messages if the RPC fails. We should just fail fast. (That logic doesn't apply to SendBatch, because
some publish errors can be continued from.)

Fixes #889.