internal/contributebot: keep PRs with "ready to submit" in sync with upstream
This is an incremental (but significant) step toward accomplishing #687.
It does not submit PRs, but it manages them so that they are more likely
to be in a mergeable state. Since this is taking some shortcuts, I am
not updating the design doc with details of this change yet, but this
proves (to me at least) the feasibility of an approach where all state
is stored in GitHub. Hopefully this eases enough maintainer toil that I
can do design work for the remaining elements on a more relaxed
schedule.

When a pull request is labeled with "ready to submit", Contribute Bot
calls syncPullRequest to update the pull request. Using a label makes
validation easy because only users with write access to a repository
can make this change.

On any branch push to the repository, Contribute Bot will search for any
open pull requests with the "ready to submit" label that have a base
branch of the updated branch. Contribute Bot will call syncPullRequest
for all the matching pull requests.

If any errors occur while merging, Contribute Bot will remove the "ready
to submit" label and leave a comment. This gives notification of
required manual intervention.

I've changed the built container to include a copy of Git. This does
bump up the container to requiring debian-slim, but I don't see a good
way around that.

Updates #687