blob: support for OpenCensus metrics
First, we combine tracing and metrics into a single package,
internal/oc.

The metrics support includes functions for creating the standard
measure and standard views.

Second, we combine tracing and metric recording into one pair of calls.
We define a Tracer type to hold common values across an API resource.

We use this machinery in blob:
- Every bucket holds a Tracer.
- We call the Tracer's Start and End methods where previously we
  used the interal/trace package.

The internal/trace package is obsolete. We'll remove it after
we replace it in the other APIs.

TODO: testing.