rabbitpubsub: increase code coverage
With a live rabbitMQ server, code coverage is over 87%.

It is lower with a fake, for the simple reason that the shims
for the live server in amqp.go are not touched. So the number
reported by our CI system is low.

Also, rework closing in the fake to remove a deadlock.