firedocstore: implement the codec test (#1547)
For Firestore, the "native" codec is implemented by the cloud.google.com/go/firestore
client. Use gRPC interceptors to obtain the protos from this client.

Add an implementation of drivertest.CodecTester, and pass it to the conformance tests.

Remove uint and complex from the test. Firestore can't handle either type.