docstore: document revisions
Support optimistic locking (safe read-modify-write cycles) by supporting
a revision field in documents.

Like an ETag, a revision field uniquely identifies a document
version. The value of the field changes every time the document is
modified.

If a document with a revision field is passed to a Docstore write
action, then the action will return a PreconditionFailed error if the
stored document's revision field doesn't match.

Since providers use different types of values for revisions, docstore
doesn't specify a single type for the field.

Implemented for firedocstore and memdocstore.

Skipping dynamodocstore so shantuo can implement.