internal/website/howto/pubsub: add section (#1704)
Like #1630 and #1669, kept as draft for now to batch how-to guides for
publication. And like #1669, the reference documentation has TODOs for
questions I couldn't answer with reference alone. I will open issues for these,
but they should not change the structure of the guides AFAIK.

As part of doing this, I updated the website's archetypes with fields I
typically add to these pages' front matter. These get used for `hugo
new`.

Updates #1576