mongodocstore: extract the ID field with a function
Previously, we assumed the key field of every collection was "_id",
the same as MongoDB's key field. That forces every collection that
wants to be portable to use that field name.

Now, have the user provide a function to extract the ID from the
document. This allows any field (or even a combination of fields)
to be used as the MongoDB _id field.