docstore: rewrite conformance tests
Add another collection to the conformance tests that is designed for
testing queries.

In order to test the full range of query behaviors for the DynamoDB
driver, we need a collection with two primary keys. To test all action
behaviors (specifically auto-generating a primary key on Create), we
need a single-key collection.

- Change the harness so it can make two different collections.
- Use the single-key collection for action tests.
- Use the two-key collection for query tests.
- Rewrite the query tests to exercise the special cases of the drivers,
  specifically use of indexes for DynamoDB and Firestore.
  (TODO: queries with multiple range clauses: see #1909.)

Other changes:

- mongo:
  - lowercase field names in queries (hack: see #1899)
  -  handle id field names other than "_id"
- firestore:
  - script for creating indexes
  - accept nil DocstoreRevision field to support structs
- dynamo:
  - script for creating tables and indexes