docstore: add Key method to driver.Collection
Add a Key method to driver.Collection that extracts the key from a document.

Use it to check that the key is always present (except for Create).

Also implement the check for ActionList that all documents must be distinct.

This is a first step in implementing the new semantics for action lists.