internal/testing/cmdtest: package for CLI testing
This package generalizes the gocdk CLI tests
(internal/cmd/gocdk/main_test.go) for use with any command-line
program.

We will use it to test our samples, and it will replace the gocdk test
framework. All that work is done and will be mailed when this PR is
submitted.

All the features here will be justified in that later work. Briefly:

- All the builtin commands are used in sample tests, and in testing
  this package itself.

- Input redirection is needed for sample sample CLI commands that
  take input from stdin.

- Variable substitution is needed in some samples to refer to
  the root directory (temporary test working directory).

- The Setup option is used by gocdk to set GOPATH.