internal/cmd/gocdk: improve module cacheability for Docker builds
Background: https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#leverage-build-cache

Before this change, the Dockerfile generated by `gocdk init` would run a
`go build` as a standalone `RUN` instruction. While this is correct, it
would typically be slow because it would have to fetch all the
dependencies on any change to the source code, since the environment
would have an empty module cache.

This change introduces a step in the Dockerfile before the `go build`: a
`go mod download` that only depends on the go.mod and go.sum. This
speeds up builds where go.mod and go.sum do not change, since the
container build environment will already have all the modules' source in
the cache. Note that this will not share build cache between runs, so
this is still not suitable for interactive builds.

The general assumption this change is making is that the Go source of
the project will change more often than the set of dependencies. While
go.mod may be out of sync with the source, there is not a huge
correctness issue, but the build will not be reproducible (`go build`
will make changes to the go.mod inside the ephemeral container).