internal/cmd/gocdk: add ECS launcher (#2416)
This adds "ecs" as a launcher type recognized by `gocdk launch`, but
doesn't add it as an option to `gocdk biome add` yet. The launcher
pushes an image to ECR then starts a task on an EC2-based ECS cluster.
I have tested this manually. The logic mostly follows the same steps
that the Cloud Run launcher does.

The launch process is effectively everything needed, but there is a
notable TODO to set up Docker auth for the ECR repository. This will be
addressed in a subsequent PR.

Updates #2301