samples/order: frontend implementation
The front-end web server for the order sample.

As I expected, writing the frontend caused me to rethink the
backend. So that has also changed: everything about the order is in
the database, and the pubsub response has just the order ID. I also
refactored order processing to make it clearer (hopefully).

Still to be done:

- Figure out how to use the response topic. Ideally this would
  cause the frontend server to signal the web page that an order
  is ready.  I'm not sure how to write that.

- Some minimal CSS to make the site presentable.