internal/cmd/gocdk/internal/launcher: create launcher that calls to a subprocess
This is primarily a proof-of-concept. The launch input is sent as stdin to the
subprocess and the subprocess must print out a valid URL on success. The process
to execute is given as the `argv` key in launch specifier, so this could also
use `go run` or the like to run an arbitrary Go "script".

As an example, the following script implements the same functionality as
`launcher.Local`:

  ```bash
  #!/bin/bash
  set -euo pipefail

  log() {
    echo "$@" 1>&2
  }

  # Save stdin to $launch_input.
  launch_input="$(cat)"
  log "Got launch input: $launch_input"

  # Convert environment variables into Docker argument list.
  mapfile -t docker_env < <(echo "$launch_input" | jq -r '.env[]')
  docker_env_args=()
  for e in "${docker_env[@]}"; do
    docker_env_args+=( "-e" "$e" )
  done
  docker_env_args+=( "-e" "PORT=8080" )

  # Start Docker image.
  docker_image="$(echo "$launch_input" | jq -r '.docker_image')"
  host_port="$(echo "$launch_input" | jq -r '.specifier.host_port')"
  container_id="$(docker run -d -p "$host_port":8080 "${docker_env_args[@]}" "$docker_image")"
  log "Container running, stop it with docker stop $container_id"

  # Print URL back to CDK.
  echo "http://localhost:$host_port/"
  ```

Updates #2376