internal/cmd/gocdk: add ECS launcher support to `biome add` (#2543)
internal/cmd/gocdk: add ECS launcher support to `biome add`

This adds resources to a ECS-launching biome to create an ECS cluster. It's
ported from the CloudFormation template that the ECS console applies when the
"Create Cluster" UI is used. Sadly, there is no convenient "single API" that
creates an ECS cluster like GKE does on GCP. Broadly, this configuration
creates a new VPC with an EC2 autoscaling group that will report back to the
ECS orchestrator. If this breaks in the future, you can create a new ECS cluster
from the AWS console, look at the resulting CloudFormation template, and make
the corresponding changes to the Terraform template.

I have manually tested this configuration and it is repeatable and results in
a working service.

Fixes #2301