mysql: Remove mysql:// prefix when passing dsn to driver
This commit removes the `mysql://` prefix before passing the DSN to the
MySQL driver for Go. The latter requires the DSN in a format like
`[username[:password]@][protocol[(address)]]/dbname[?param1=value1&...&paramN=valueN]`
as described in
https://github.com/go-sql-driver/mysql/blob/master/README.md#dsn-data-source-name.

It doesn't accept a `mysql://` prefix and, when used, will not parse the
DSN correctly.

Fixes #2644