secrets/hashivault: Add support for VAULT_ADDR and VAULT_TOKEN
Fixes: #2990

`VAULT_ADDR` and `VAULT_TOKEN` are the recommended way of configuring
access to vault as per the Vault docs https://www.vaultproject.io/docs/commands#environment-variables

This commit adds support for those environment variables. We check
that the *existing* VAULT_SERVER_URL and VAULT_SERVER_TOKEN env vars
are not set first. If they are, then we use those. If they are not
then we fall back to the new environment variables