pubsub/kafkapubsub: Configure initial offset in URL.
This commit adds the ability for users of the `kafkapubsub` package to specify the initial consumer offset
when using a URL.

When creating a kafka consumer, you sometimes want to specify at which point the consumer should begin reading from
a topic. This is specified as either `oldest` or `newest`, where `oldest` is the oldest committed offset by the
consumer group and `newest` is the most recent offset.

Currently, there is no way to configure this via the URL, you'd have to use the `kafkapubsub` package directly to
build your consumer. So when you use a URL, you get the default sarama behaviour and you start consuming from
the newest point in the stream. This doesn't allow you to read streams from the start, which is a very common
usecase for consumers.

Signed-off-by: David Bond <davidsbond93@gmail.com>