Provide mechanism for sending ClientId
Due to Azure security requirements, there might be multiple service
identities attached to a particular account. In that case, the Client ID
must be provided in order to secure an authorization token.

This allows the `AZURE_CLIENT_ID` environment variable to be read in as
an optional extra bit of data and if provided will invoke the
appropriate underlying method in Azure autorest to authenticate with the
client Id instead