gcp/cloudsql: enable IAM login
This enables the IAM login option for GCP, which is backwards-compatible
as other forms of authentication (such as username/password) continue to
function as expected. We could also have this configurable through a URL
parameter, but that seems like the wrong approach. If you truly for some
reason didn't want IAM authentication for cloud sql, you would have
already declined to enable the flag it in the GCP setup for that
database instance. If you did want to use IAM authentication, it would
be annoyingly hard to debug why it wasn't working as it may surprise you
to learn that the proxy has it off by default. So, it makes the most
sense to just always enable it in the proxy.

Fixes #3121.