docstore/mongodocstore: fix error when calling the update function with no 'set' operations
This update modifies the newUpdateDoc function in the docstore/mongodocstore package to check the length of the 'sets' variable before adding the '$set' field to the 'updateDoc' map. This prevents an error when calling the update function with only 'unset' or 'inc' modifiers and no 'set' operations. This change improves the flexibility and robustness of the update function.

Resolves #3393