Update Mongo dialer when MONGO_SERVER_URL rotates
Prior to this commit, the dialer for MongoDB was generated once from
MONGO_SERVER_URL environment variable but was never updated even when
the environment variable was updated in subsequent calls. While this
works fine when MONGO_SERVER_URL is not expected to update, but as
MONGO_SERVER_URL also contains the credentials to connect to MongoDB,
it's a fairly common use case to rotate these credentials (and hence the
environment variable) at regular intervals.

This commit fixes that and updates the dialer when MONGO_SERVER_URL is
updated.