s3blob: fix data race
w.pr cannot be used to indicate that the pipe has been closed. The problem with
using w.pr is that a write to it in the goroutine launched by writer.open can
race with a read on w.pr in Close.

We just don't write to w.pr in writer.open. The read end of the pipe can still be
closed twice, but that's fine.